backend/requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1.18
djangorestframework>=3.15
django-cors-headers>=4.4
PyJWT>=2.9
djangorestframework-simplejwt[crypto]>=5.3
channels>=4.1
channels-redis>=4.2
celery>=5.3
redis>=5.0
reportlab>=4.2
pdfkit>=1.0.0
Pillow>=10.3
cloudinary>=1.41
python-dotenv>=1.0
openpyxl>=3.1
